Move bundle extensions to shared
