+++ macOS/LocalPackages/DataBrokerProtection/Sources/DataBrokerProtection/macOS/Bundle/DataBrokerProtectionBundleExtension.swift
@@ -1,7 +1,7 @@
 //
-//  BundleExtension.swift
+//  DataBrokerProtectionBundleExtension.swift
 //
-//  Copyright © 2023 DuckDuckGo. All rights reserved.
+//  Copyright © 2024 DuckDuckGo. All rights reserved.
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -18,16 +18,17 @@
 
 import Foundation
 
-protocol GroupNameProviding {
-    var appGroupName: String { get }
+public extension UserDefaults {
+    static let dbp = UserDefaults(suiteName: Bundle.main.dbpAppGroup)!
+    static let config = UserDefaults(suiteName: Bundle.main.configAppGroup)!
 }
 
-extension Bundle: GroupNameProviding {
+extension Bundle {
 
     static let dbpAppGroupName = "DBP_APP_GROUP"
     static let configAppGroupName = "APP_CONFIGURATION_APP_GROUP"
 
-    var appGroupName: String {
+    public var appGroupName: String {
         guard let appGroup = object(forInfoDictionaryKey: Bundle.dbpAppGroupName) as? String else {
             fatalError("Info.plist is missing \(Bundle.dbpAppGroupName)")
         }
@@ -35,6 +36,20 @@ extension Bundle: GroupNameProviding {
     }
 
     var configAppGroupName: String {
+        guard let appGroup = object(forInfoDictionaryKey: Bundle.configAppGroupName) as? String else {
+            fatalError("Info.plist is missing \(Bundle.configAppGroupName)")
+        }
+        return appGroup
+    }
+
+    var dbpAppGroup: String {
+        guard let appGroup = object(forInfoDictionaryKey: Bundle.dbpAppGroupName) as? String else {
+            fatalError("Info.plist is missing \(appGroupName)")
+        }
+        return appGroup
+    }
+
+    var configAppGroup: String {
         guard let appGroup = object(forInfoDictionaryKey: Bundle.configAppGroupName) as? String else {
             fatalError("Info.plist is missing \(Bundle.configAppGroupName)")
         }
